requirements.txt
aiomysql
tortoise-orm
fastapi
pydantic
python-dotenv
uvicorn
python-jose
passlib
python-multipart
bcrypt